requirement.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.1
django-cors-headers==3.1.0
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
DjangoUeditor==1.8.143
idna==2.8
Markdown==3.1.1
mysqlclient==1.4.4
Pillow==6.1.0
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
